_examples: add imagebutton widget example
This is a current way to implement a simple
image button with ebitenui.

Not to be confused with TextureButton which
usually means "no button outlines, just a sprite".
An imagebutton is something that has both.
Imagine an inventory where every slot is a button
holding an image of an item. Or a menu where you
choose which units to build: every button is pressable
and has unit icon as an image.