Remove unused error prone key just pressed helper
All widgets use IsKeyPressed and an internal boolean state to store the 'just' side of it. The reason for this is that there may be e.g. two ebiten updates and only one rendering, therefore the keys are never just pressed when tested in the render method of widgets.

We might as well delete the IsKeyJustPressed input helper function given it cannot be relied upon.