widget: make button text-related options more granular
This patch introduces an intermediate configuration object
those fields are updated by the user-provided options.

When all user-provided options are executed, there is
a new initialization step called initText that uses
the configuration object to lazily build the text component.

This approach can be used in most (all?) widgets that
may benefit from easy-to-override styles.