fix Text widget width calculation in `measure()`
The bug was something like an off-by-one mistake:
an extra space width was added to the Text's width
if BBCodes were enabled.

Also avoid calculating the word width if we're about
to discard that result when removing the BB-code tags.
Instead, check for the tags first and then calculate
the width one way or another.